requirements.txt
opencv-python==4.8.0.76
Pillow==10.0.0
RPi.GPIO==0.7.1
pynput==1.7.6
imutils==0.5.4
pyserial==3.5